Add param scanner, inprogress paramSpace plot
alternate_param_scanner is designed to create a slice along a new parameter at a point on an already created branch

plot_paramSpace is a plotting program which can make a plot based on user defined parameter changes. Relevant to the scanner so you can see first and second continuation params on a single plot.