m_sys_cond_RWFold, m_sys_rhs_RWFold work with single 'rho', designed for arbitrary numbers of rho
m_sys_cond_RWFold will support ind_rho as a double array.
	Designed for an arbitrary amount of rhos.
	Tested with a single rho.

m_sys_rhs_RWFold will support the same as above, and is tested as above.

--

I will now need to make all the rest of the fold files work with ind_rho as a 'double' array.

This is how I plan on implementing multiple rhos. Adding multiple rhos to the end of
point.parameter, but saving the index for each in ind_rho.